# docs/swagger.yaml
basePath: /api
definitions:
...
  models.PostTag:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      password:
        description: Será armazenado como hash
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
host: localhost:8080
info:
  contact:
    email: seuemail@example.com
    name: Seu Nome
    url: http://www.seusite.com
  description: API para gerenciar posts, comentários, categorias e usuários.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
...
  /tags/{id}:
    delete:
      consumes:
      - application/json
      description: Deletar uma tag por ID
      parameters:
      - description: ID da tag
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag deletada com sucesso
          schema:
            type: string
      summary: Deletar uma tag por ID
      tags:
      - Tags
    get:
      consumes:
      - application/json
      description: Buscar tag por ID
      parameters:
      - description: ID da tag
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.PostTag'
      summary: Buscar tag por ID
      tags:
      - Tags
    put:
      consumes:
      - application/json
      description: Atualizar uma tag por ID
      parameters:
      - description: ID da tag
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Tag atualizada com sucesso
          schema:
            type: string
      summary: Atualizar uma tag por ID
      tags:
      - Tags
...
  /users:
    get:
      consumes:
      - application/json
      description: Listar todos os usuários
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Listar todos os usuários
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: Criar um novo usuário
      parameters:
      - description: Usuário a ser criado
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Criar um novo usuário
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Deletar um usuário por ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Usuário deletado com sucesso
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Deletar um usuário por ID
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Buscar um usuário pelo ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
      summary: Buscar um usuário pelo ID
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Atualizar um usuário por ID
      parameters:
      - description: ID do usuário
        in: path
        name: id
        required: true
        type: string
      - description: Usuário a ser atualizado
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: Usuário atualizado com sucesso
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Atualizar um usuário por ID
      tags:
      - Users
...
swagger: "2.0"